Refactor: moving `system_tray_icon` to separate module.
