+++ app/ui/settings_dialog.py
@@ -1,8 +1,8 @@
-from PySide6 import QtWidgets, QtGui
+from PySide6 import QtGui, QtWidgets
+from utils import get_asset
 
 from app.hd_active import HdActive
 from app.ui.forms.settings_ui import Ui_Dialog
-from utils import get_asset
 
 
 class SettingsDialog(QtWidgets.QDialog):
